requirements.txt
streamlit>=1.29.0
Pillow>=10.1.0
bcrypt>=4.1.2
requests>=2.31.0
python-dotenv>=1.0.0